feat(FN-1616): move view model types to correct directory
