+++ trade-finance-manager-ui/server/controllers/utilisation-reports/utilisation-report-reconciliation-for-report/index.ts
@@ -4,6 +4,10 @@ import api from '../../../api';
 import { asUserSession } from '../../../helpers/express-session';
 import { PRIMARY_NAVIGATION_KEYS } from '../../../constants';
 import { mapFeeRecordItemsToFeeRecordViewModelItems } from '../helpers';
+import { UtilisationReportReconciliationForReportViewModel } from '../../../types/view-models';
+
+const renderUtilisationReportReconciliationForReport = (res: Response, viewModel: UtilisationReportReconciliationForReportViewModel) =>
+  res.render('utilisation-reports/utilisation-report-reconciliation-for-report.njk', viewModel);
 
 export const getUtilisationReportReconciliationByReportId = async (req: Request, res: Response) => {
   const { userToken, user } = asUserSession(req.session);
@@ -16,7 +20,7 @@ export const getUtilisationReportReconciliationByReportId = async (req: Request,
 
     const feeRecordViewModel = mapFeeRecordItemsToFeeRecordViewModelItems(utilisationReportReconciliationDetails.feeRecords);
 
-    return res.render('utilisation-reports/utilisation-report-reconciliation-for-report.njk', {
+    return renderUtilisationReportReconciliationForReport(res, {
       user,
       activePrimaryNavigation: PRIMARY_NAVIGATION_KEYS.UTILISATION_REPORTS,
       bank: utilisationReportReconciliationDetails.bank,
