+++ trade-finance-manager-ui/server/types/view-models/index.ts
@@ -1,2 +1,3 @@
 export * from './find-utilisation-reports-by-year-view-model';
 export * from './previous-bank-reports-by-year-view-model';
+export * from './utilisation-report-reconciliation-for-report-view-model';
